Pull cost goals together
