+++ touchdown/aws/__init__.py
@@ -17,6 +17,7 @@ from . import (
     vpc,
     ec2,
     cloudfront,
+    cost,
     elasticache,
     elastictranscoder,
     elb,
@@ -36,6 +37,7 @@ __all__ = [
     'vpc',
     'ec2',
     'cloudfront',
+    'cost',
     'elasticache',
     'elastictranscoder',
     'elb',
