+++ touchdown/aws/cost/rds.py
@@ -12,66 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# This code is not currently exposed publically. It is an example of how to
-# stream from a aws log using the FilterLogEvents API.
-
-import requests
-import re
-import demjson
-
 from touchdown.core import plan
 from touchdown.aws import common
 from touchdown.aws.rds import Database
 
-
-class PricingData(object):
-
-    REGIONS = {
-        "ap-northeast-1": "apac-tokyo",
-        "ap-southeast-1": "apac-sin",
-        "ap-southeast-2": "apac-syd",
-        "eu-central-1": "eu-central-1",
-        "eu-west-1": "eu-ireland",
-        "us-east-1": "us-east",
-        "us-gov-west-1": "us-gov-west-1",
-        "us-west-1": "us-west",
-        "us-west-2": "us-west-2",
-        "sa-east-1": "sa-east-1",
-    }
-
-    def __init__(self, url, **tags):
-        self.url = url
-        self.tags = tags
-
-    def get(self):
-        data = requests.get(self.url).content
-        data = re.sub(re.compile(r'/\*.*\*/\n', re.DOTALL), '', data)
-        data = re.sub(r'^callback\(', '', data)
-        data = re.sub(r'\);*$', '', data)
-        return demjson.decode(data)
-
-    def get_region(self, region):
-        filtered = filter(
-            lambda r: r['region'] == self.REGIONS[region],
-            self.get()['config']['regions'],
-        )
-        return filtered[0]
-
-    def get_instance_type(self, region, instance_type):
-        r = self.get_region(region)
-        for i in r['types']:
-            for t in i['tiers']:
-                if t['name'] == instance_type:
-                    return {
-                        "currency": "USD",
-                        "cost": t['prices']['USD'],
-                    }
-
-    def matches(self, resource):
-        for tag, value in self.tags.items():
-            if getattr(resource, tag, None) != value:
-                return False
-        return True
+from .common import PricingData
 
 
 pricing_data = [
